Added a top-level release script
The release script will create releases and move them to the
system-release directory. It will also optionally create tarballs of the
available releases. One can also specify a particular project to release
such as main of a ext_rel project such as web_frontend.

This also removes the dummy_app from the main release since that has
been removed from the repo. It also extends the .gitignore to ignore
crashdumps and the system-release dir.

Change-Id: I946b533447ab022abc442c1e260d77ea2d6e3c8b