[#222] add initial game app
the game application is now up and running.
there is a game release in ./rel

the operation new_game is supported. [asynchronous]
It uses the newly created game-record in the datatypes library.
There was a small modification in the db app to allow the atom
'undefined' for keys when writing to the database.
A key is generated in that case.

Patch: fix whitespace problems

Change-Id: I7e5b2cf18585597215de82854605016e5e01b796