[#203] Add echo XMPP component and test this at build.
The component is added to the xmpp_frontend and built
with the rest of the frontend.

The build script starts the newly-built ejabberd with
the echo component installed, runs eunit tests that check
that it works, and stops ejabberd again.

Downloading libexpat has been removed since libexpat
is now a requirement for machines where this should
be built.

Change-Id: Ifbbd5663ce149a92f80f3bf0307ceb55b1587841