[#131] web_frontend now net_adm:pings the backend during startup
It will fail, if it is started and there is now backend running.
It produces a readable error message. Look for it.

fix a bug in user_record_gen - used to store everything as binary

create_user now works for the first time as well, pg2:start_link() in
the app startup was necessary.

Change-Id: I3c2de6e8922c2acb019d85ba78836582bec99e05