[#271] first steps for the rule engine.
get_reachable functions work.
Basic unit support (add_unit, remove_unit, move_unit, get_units) in map_utils.
Basic order validity checks in rules.
The digraph_io module allows to convert a map to an erlang term and reconstruct
a map from that term. This is necessary for saving graphs to the database.

gave up on including dialyzer in the build script because the plt makes problems
on server3 (and pbly on your machines, too), there is make dia, though

Change-Id: If3e371240a879e5c03bbe6b0c87f606174b585f3