[#324] Changed the DB wrapper into a service.
Each riak node has a limited amount of supported connections,
therefor we can't have one connection for every worker in our
backend services. The solution for this is that we have the DB
as a service itself with workers. The db workers perform the
requests from the other service workers and abstract away the
DB connection. Now only DB has active connections to riak.
=> we don't reach the connection limit

Change-Id: I5b5dd88e3b3456c31ce2803139cab8544920c8a9