[#271] Add unit-dict, create map_data module, rename map_utils to map, write map:get_provinces
The unit dict is necessary for conflict resolution (we need to know, which
unit was moved this round and which wasn't).

map_data was created because the data is rule-specific, not map-specific and
because the map_utils was growing fast.

map_utils contains the full map and was therefore renamed to map.

Change-Id: I5e417f5f71c231b87c5460fe68735da3e2e80276