[#271] Create map_data module, rename map_utils to map,
write map:get_provinces, work on tests, add coverage to rebar.config, add unit_can_act feature and test coverage reporting

The newly added unit dict is necessary for conflict resolution (we need to know, which unit was moved this round and which wasn't).

The convention for the dicts is: if a key/value was not set, a get operation
returns the atom undefined.

In the case of get_province_info(Map, Province, center),
check for true/undefined as opposed to true/false.

map_data was created because the data is rule-specific, not map-specific and
because the map_utils was growing fast.

map_utils contains the full map and was therefore renamed to map.

rules:unit_can_act checks whether a unit is free to act or has moved this round
already. This includes adding some notion of "time" which is not visible to the outside.

Test coverage is now measured and integrated in the documentation.

Change-Id: I8b83194338bd9cc2b290444118e6e4ad825fd735