Add call and cast concepts to xmpp messages via xmpp test client.
Also enable old tests.

Now there's a xmpp_call function in xmpp_client which sends the
provided message to the provided user, and returns the first
message received by the xmpp test client. This means we don't
wait and hope that the message last received message after
x seconds is the one that we wanted to receive, and we only
wait as long as we need to. No response means a test timeout
or a gen_server timeout, both of which will fail tests.

Removed empty message test. tt_bot ignores empty messages because
they have no body tag, and messages with no body tag can be any of
a big range of info messages, e.g. "user started writing" or "user
paused". The test for the server doing nothing is...wait for timeout.
This would only tell us that the server didn't respond within the
timeout. It's not particularly useful any more, even though it was
quite important when in the days of tt_register_bot.

Change-Id: I9f2bfa656f15d9f1f8570e49c6a9215719f352e4