[#271] Implement basic hold and move orders.
A lot of refactoring was done, rules.erl will loose Diplomacy-specifics in
the near future, rules will be implemented in diplomacy_rules.erl and
only executed by rules.erl. I am using a kind of strategy pattern here.

Partial Orderings of rules will be a tough one, but a naive implementation
would be possible, too.

To the reviewer: the main change lies in the way we define rules now.
They are implemented in diplomacy_rules.erl.

Change-Id: Ife96508202c6fe0e1d26066c22f276422b6b15b4