[#283] Game timer that changes according to given data
A game timer that changes phases according to the data
to which game it belongs to.
The game timer is started when a new game is created and
starts to run immediately.

When a phase change occurs the game is called to handle it.
The game timer is updated according to changes in the game,
if it's still in the waiting phase.

Change-Id: I3c96a29ea48b9d40f5fe24491d1268d1ea194a8b