[#271] test diagrams no.7-12 make them pass
Map has a get_units/1 function now, that returns [{Province, Unit}] for all
units on the board.

I added reply functionality to the rule's possibilities, they now can reply an
erlang term that will end up in the output of rules:process.
This is needed for dislodging orders in this special case but could be used for
verbose justifications of errors, too.

diplomacy_rules.erl now lives up to the fact that support orders are only valid
for the specific orders they are supporting but not the supported unit in
general. This is implemented by adding the whole support order to the
unit-dict's support_orders field.

Change-Id: I04371a2e7e4b7fe76ac86d1e2e9021513d8fbdad

change support system, it's tracking the whole support orders now

Change-Id: I077cc342b429712f87f1a80d4bca1fb54a52d857