Ensure that the get-deps script finds all deps
By building riakc and the db app, we ensure that all dependencies can be
downloaded by get-deps properly.

This also fixes the lib path in smtp_frontend and adds db as a dependency
for both smtp_frontend and xmpp_frontend. This is not nice, but it is
needed so that all dependencies can be downloaded up front, and not while
building.

Change-Id: I5e2adcb724d654833c6682417c813405e90e13fb