Make build and test use top-level rebar file
Instead of running rebar compile in each app dir, the build script will
use a top-level rebar config file to drill down in the app directories.
This speeds up the build quite a lot and is a much nicer solution. The
same goes for testing which now calls rebar eunit in the root dir.

All local build/test scripts should still work, but they are not called
when making a top-level build. By setting lib_dirs and deps_dir in a
consistent way, we can keep the old behaviour of local scripts.

This also changes the names of release .tar.gz files and redirects some
output from wget to stdout instead of stderr to make Buildbot happier.

Change-Id: I38c7b48170d477b14c50d56b128255e9d7cdb10e