[#343] Use rebarized ejabberd for xmpp_frontend
This change removes the xmpp_frontend in ext_rel and adds a xmpp_lib in
apps. When one releases xmpp_frontend, the compiled files in xmpp_lib
will be included in the release. The xmpp_frontend is basically ejabberd
with our components added.

The new xmpp_frontend release is invoked by starting xmpp_frontend
instead of ejabberd. Ejabberdctl works as expected.

Due to reltool, exmpp must be excluded from all releases to not clash
with ejabberd.

Change-Id: Ic6bc3a763ab11d130ba729e172b65b710121cb16