[#342] Only release what's needed + new Makefile
This remove the releasing of db, game and user_management. Releases are
now done directly to system-release with rebar generate. This also adds
a top-level Makefile for building, testing and releasing the system.
The intent is to remove the old build, test and release scripts in the
future.

The integration tests are now run through the ext_test/test script only.

A stupid missing semicolon is added to endpoint.yaws in web_frontend.

Change-Id: Id3b0fba557c934e37c750c6fe55969e6c2bf88ce