Clean up game timer when game stops or finishes
There is still not a notion of a "finished" game from the
rules engine.
The game timer will change the state of the game record to the
new given state, it will also unregister its name and die gracefully.

A game timer that had a normal termination will not be restarted,
but the supervisor will restart if the termination was unexpected,
the game timer does not handle restarts yet.

Change-Id: Ifd0098df509bc38f6e9c42a5f6fb51cc8a736ab9