Move name of node to separate config file
Since we want to be able to set the node name in a more dynamic fashion,
move its setting to a special file called nodename. Adapt the startup
scripts to use that file and add it to reltool.config. Also remove the
Windows batch files since we don't support that platform.

Change-Id: I75ecbab28336cbce9b644f37af1ca41787585c86