[#357] Add skeleton of system manager application
This adds a system manager app that is supposed to be run in a separate
VM and to be used for configuring the system. The idea is that one can
connect to the system manager and dynamically change configuration of a
server and also start and stop releases. This is only a skeleton without
any functional code.

Change-Id: I497871b36b494cbcf9152dfb26098a4dd6470720