add replication of session table
when a new backend is added that finds a list tagged with backend_nodes in it's
config, it will try to ping them. The first one that reacts will be used as
a source to replicate the session-table from. If there is no such list or it is
empty or no backend is ping-able, the backend will create a session table as
it did before that change. This table can then be replicated by subsequent
backends.

Change-Id: If851b29054113e558095978b1c9e75ae16d2b33e