[#357] Add module for starting/stopping releases
This adds a handle_releases module that allows one to start, stop and
ping releases by running their control scripts.

The startup scripts for releases is modified to first look for a file
called app-override.config in etc/ before using the regular app.config.
This is a preparation for enabling system_manager to override app files
in a sane way.

Change-Id: I105df3b3c0dc2308d54717ff97ae0f73164fdbce