[#359] Necromancer - first working version
handling of corpses works for session_presence, the session is deleted from
mnesia by a random other node when the node that wrote it dies.
restarting of games works, too, BUT that's not perfect yet:
a) games are not saved as long as they are in the waiting phase
b) when games are restarted because of a node going down, they do not compensate
   for the already-spent waiting time: they wait the whole phase waiting time
   again.

Problem b) is, IMHO, acceptable in our context.

This commit is not fully ready, but I won't be here tomorrow and therefore
I am committing this now.

There are *NO* automated tests. I think that should wait for the system manager
app.

Dialyzer -- I made some modules warning-free, for instance:
* controller,
* db
* necromancer
Break the specs and I WILL come to haunt you in your dreams, Freddy Kruger
style.

Change-Id: I6a1df0558eda1fe3549e0016d07586b7c2f792fa