Add support for adding Riak to our releases
Since there are a system manager now that can configure Riak for us, it
is possible to distribute Riak in our debian package. This adds support
for building and releasing Riak to system-release as a part of the
dependency file generation/download. This is not very nice but should
work well in practice. This solution avoids building and releasing Riak
with our own code, which is nice.

Change-Id: Ide61a7f18442053c3baf238d279183f8ef11f93a