[#84][#93][#94][#393] Moderator/Operator game player communication
A moderator or operator can talk to all or a particular player of
a game, given all or a part of the countries and a game id.
They cannot communicate with game players if the game has not started.

[#392] Operators/moderators can send these kind of messages through all
interfaces.

This commit also includes some bugfixes in the web chat interface.
Change-Id: Ib69411b9cb4b6d1e5fd78bd1b7a37aa1f969d839