[#357] Make cluster_manager script more friendly
This adds getopt as a dependency and uses it for parsing command line
arguments to cluster_manager. It also adds the ability to stop releases
which can be useful sometimes. A small bug fix is also included:
"controller" should be "controller_app" in the backend config file.

Escriptize is only done for cluster_manager now and not for its
dependencies.

Also update the backend app.config to actually match the variable names
in db.app.src.

Change-Id: I6b69617ec3a92a864c9a49c2c606c50de873b1bb