[#397] Load test scaling script
scaling_test.sh test how the system scales.
It runs the provided configurations and adjusts the load
in the basho config, example:
$ ./scaling_test.sh 1 2 4
Will run a test with an equal system on 1 machine with 1
concurrent basho worker. Afterwards the same with 2 machines
and 2 load workers, and so on...
An equal system is 1 backend node and 1 riak node on one machine.

* Updated the general test to be more random.
  Otherwise all the load ends up on 1 machine.

Change-Id: I8b55ac0741d3dc478c82c4a10872d8cab6923d81