[#398] Restart Game Timer and continue Phase
Game timer will pick up from the phase it was, it will not take
into consideration how much time it has spent in that phase, but
wait for the full time of the phase before moving on to the next.
Games that have not been started (are in waiting phase) are not
handled.

Change-Id: I196ce9c963f6bbc71e0fb29bafa6589ca35bb0a0