Rework releases to only create one package
Creating serveral release packages, one for each release, is a waste and
creates duplication. This change removes the generation of several
release packages in favor for a single one. This package includes all
our previous releases and releases are started and configured as before,
but with a single path prefix "system-release/tt" instead of
"system-release/backend" etc.

The system manager and integration tests has been updated to reflect
this new release structure.

Change-Id: I10476470c54edbdab95b81bb12174de15416dc9d