[#446] More flexible version of the tt_general flow
Each operation in the tt_general flow is now defined as a
basho bench driver operation. Meaning the frequency of
operations in relation to others can be defined.

Therefore the rate of operations can be used in the config. ({mode, {rate, X}}.)

By adjusting the frequency of operations to something "real" and
setting the rate to a realistic value we can find the max amount of users
for our system by varying the amount of load workers ("concurrent").

Change-Id: I7a175a442126ffe4a3a5b0118e2d2450f956a549