[#413] change put-props
changed most of them to {w,1}, some of them where converted to a
{w,0} (where the return is not necessary, eg. for message logging).
In those cases, a timer:sleep(100) was inserted in the relevant tests so
that the data can propagate before the values are read again.

Ran the new optimized version and compared the new results vs the old results.

Change-Id: I6077463fd5002d8d82c8ada1d8c2b61c5cfcd2d0