Make our tests use sm/cm and local riak
This change modifies the way both unit tests and integration tests run.
Since we now have very good support for starting and stopping releases
in the cluster manager, we should use it. This changes our tests to
start all neccessary releases in that way. It also uses a local riak
install for tests.

In other news, the startup scripts for releases are tweaked so that they now
wait for a certain application to start before returning. This does not
guarantee that everything actually started up in a good state, but at least one
knows that applications has been loaded before continuing.

There is also a small change in the XMPP frontend tests to use macros
for setting the cookie and backend to connect to during tests.

Change-Id: I07698eaec3f34b24ac2100630ce00d752d4ad574