Make system manager use workers to handle calls
This makes the system manager use the same worker framework as the rest of the
server applications. This is useful for starting and stopping releases in
parallel from the cluster manager. Beware that it is up to the caller (cluster
manager) to ensure a proper serialization of starting and stopping releases.

This change also lowers the memory usage of our releases with about 500 Mb in
total when running them all. This was achieved by lowering the maxiumum number
of ports/sockets available from 250k to 100k.

Change-Id: Ibcd76b657dc79ae7fe70061520ca51ae621f8943