Load test results: finding max requests/sec
In this load test we tried to find the max requests/sec the system can handle
with different setups. Therefore we ran a bunch of tests that generated
different load to see when it goes down.

Unfortunateley the used load was still too high.
On the plus side, now we know in which area to search in the future.

Change-Id: Iecf85d3aead59b56d6ca321b9e37c3e3d4b8eab6