Fix race condition in system manager
When doing parallel start/stop in the system manager, a race condition
could occur when writing statuses of releases to file. This
functionality has been refactored into a separate gen_server module that
ensures serialized access to the file.

This change breaks the abstraction in the system manager in one place:
when terminating the manager, the release_status_tracker worker will
stop and started releases instead of the system_manager_worker. This is
because of that the system_manager_worker uses the service library where
one cannot easily change the child timeout on shutdown.

Change-Id: Ied3d7ce7908e4b0d8ee029a57a98b4ec2973258e