[#402] Add test for node failing when game is in a particular phase.
This will run one long, slow test without messing with the internals
of the backends to be as realistic as possible.
It will then run for each of the ongoing phases the same test,
but uses RPC to find out where the game is, advance the game to the
phase we want to test using the sync_event hack, kill that backend,
and then wait to see if the game still advances automatically.

Also:
- Make game_join_proc write with w,1 so that timer reads updated value.

Change-Id: I81fb0ee4d95699f87d57226f09160753c88aa5ca