Phase change pushing
web: a player in a game is notified with a message (top screen)
when the phase is changed for a player he/she is playing, if
the player is viewing the game, the full overview is updated.

im/mail: the user receives a message with information about
phase change only.

Change-Id: I6535684f226880423623df22d5527085ba780323