Cleanup of rebar.config and app.src files
This is an attempt to clean up both rebar.config files and app.src files so that
they are more consistent. Stuff that is not needed any more has been removed and
all applications has real descriptions now. This adds lager as a build
dependency for almost all applications since we should use it more fore logging
in the future.

Change-Id: I17dbef0ae4fb406b0dad77674899251eb9d9c8a6