Cleanup how external tests are compiled and run
This change renames ejabberd_echo to xmpp_integration_test and removes several
files that is no longer needed there.

An updated version of Rebar is also provided since this gives us the ability to
use the apps=/skip_apps= arguments. This new functionality is used to move the
compile step of our integration tests into the top-level rebar.config. It also
simplifies how we call integration tests in the test script.

Change-Id: I0318acbbf66d29dfe796d2346ed39714fee1c3e2